refactor: move autocomplete into its own component
